feat: Implement deletePost and updatePost functions in apiService.js
The code changes in apiService.js add the deletePost and updatePost functions to handle deleting and updating posts respectively. These functions are used to make API requests to delete a post and update a post's content. This enhances the functionality of the application by allowing users to delete and update their posts.